+++ tf-modules/archive/variables.tf
@@ -397,3 +397,9 @@ variable "dead_letter_recovery_memory" {
   default = 1024
   description = "The amount of memory in MB to reserve for the dead letter recovery Async Operation Fargate Task"
 }
+
+variable "enable_otel_tracing" {
+  description = "Enable OpenTelemetry tracing and X-Ray"
+  type        = bool
+  default     = false
+}